Cancel mousedown and touchmove
On desktop, we want to cancel mousedown, so the user doesn’t start
scrolling the window or highlighting text while moving their mouse
around.

On mobile, if we cancel touchstart, the subsequent click event doesn’t
fire.  In that case, we don’t want to cancel touchstart, but instead
cancel touchmove.