Replaced option by a feature detection
It's inspired by https://gist.github.com/lorenzopolidori/3794226
and relies on getComputedStyle(). I chose to sandbox the check in an iframe, because it seems it may crash android browser if done on the document directly.

Following the way you built ./classes.js, I chose to externalize the styling logic in a styleProperty module, that exposes `check` and `set` methods.