Expose the spill event
When the `removeOnSpill` option is set to `true`, a new event called `remove` is exposed that is not otherwise available. However, as the name of the option indicates, the `remove` event is really just a side effect of a `spill` event, but the spill event itself is never emitted. This PR exposes the spill event when `removeOnSpill` is set to `false`, so application logic can hook into the same set of conditions that would otherwise have fired a `remove` event.