Feature: mark an element to do a horizontally search for a container if pointer is over that element
Use that if you have an element that is sibling of a container element
(e.g. title of the container) and you want to connect that element to
the container.
In this case the dragged item will be automatically appended to the
container if the pointer is over that sibling element.

Set that property with a class name that is also set to the sibling
elements you want to connect with their sibling container.