fix: allow calling release without argument
Calling `release({})` used to be okay. The code of `release` would try
to grab `clientX/Y` from the empty object, get `undefined` for both
coordinate, and feed them to `elementFromPoint`. `elementFromPoint`
was fine with that and just returned `null`. However a recent Chrome
change broke this. Chrome now checks that the coordinates passed to
`elementFromPoint` are finite doubles and so passing `undefined` no
longer works.

This commit makes it so that calling release without an event object
works. It will just skip trying to find a drop target.