Mirror image issues
Fixed a performance issue occurring when dragging an element that contains child elements. A mirror image was being created for each child image as well as the containing element. This lead to performance issues when starting to drag and waiting for elements to settle, so that the element could be dropped. This fix just checks if a mirror image already exists, and if so, doesn't append another one.